Refactor scalacheck instances
