+++ modules/scalacheck/src/main/scala/ahlers/faker/scalacheck/company/implicits/CompanyArbitraryInstances.scala
@@ -1,0 +1,9 @@
+package ahlers.faker.scalacheck.company
+package implicits
+
+import ahlers.faker.Company
+import org.scalacheck.Arbitrary
+
+trait CompanyArbitraryInstances {
+  implicit val arbCompany: Arbitrary[Company] = Arbitrary(genCompany)
+}
