+++ modules/scalacheck/src/main/scala/ahlers/faker/scalacheck/company/implicits/package.scala
@@ -1,0 +1,3 @@
+package ahlers.faker.scalacheck.company
+
+package object implicits extends CompanyArbitraryInstances
