+++ modules/scalacheck/src/main/scala/ahlers/faker/scalacheck/company/package.scala
@@ -1,0 +1,8 @@
+package ahlers.faker.scalacheck
+
+import ahlers.faker.Company
+import org.scalacheck.Gen
+
+package object company {
+  val genCompany: Gen[Company] = opendata500.genCompany
+}
