+++ modules/scalacheck/src/main/scala/ahlers/faker/scalacheck/implicits/package.scala
@@ -1,0 +1,6 @@
+package ahlers.faker.scalacheck
+
+import ahlers.faker.scalacheck.company.implicits.CompanyArbitraryInstances
+import ahlers.faker.scalacheck.name.implicits.NameArbitraryInstances
+
+package object implicits extends CompanyArbitraryInstances with NameArbitraryInstances
