+++ modules/scalacheck/src/main/scala/ahlers/faker/scalacheck/name/census1990.scala
@@ -1,16 +1,18 @@
-package ahlers.faker.scalacheck
+package ahlers.faker.scalacheck.name
 
-import ahlers.faker._
-import ahlers.faker.datasets.census1990._
-import org.scalacheck.Gen._
-import org.scalacheck._
+import ahlers.faker.datasets.census1990.FamilyNamesIterator
+import ahlers.faker.datasets.census1990.GivenNamesFemaleIterator
+import ahlers.faker.datasets.census1990.GivenNamesMaleIterator
+import ahlers.faker.PersonFamilyName
+import ahlers.faker.PersonGivenName
+import org.scalacheck.Gen
+import org.scalacheck.Gen.oneOf
 
 /**
  * @author <a href="mailto:michael@ahlers.consulting">Michael Ahlers</a>
  * @since May 18, 2020
  */
 object census1990 {
-
   val genPersonGivenNameFemale: Gen[PersonGivenName] =
     oneOf(
       GivenNamesFemaleIterator()
