+++ modules/scalacheck/src/main/scala/ahlers/faker/scalacheck/name/implicits/census1990.scala
@@ -1,8 +1,10 @@
-package ahlers.faker.scalacheck.implicits
+package ahlers.faker.scalacheck.name.implicits
 
-import ahlers.faker._
-import ahlers.faker.scalacheck.census1990._
-import org.scalacheck._
+import ahlers.faker.scalacheck.name.census1990.genPersonFamilyName
+import ahlers.faker.scalacheck.name.census1990.genPersonGivenName
+import ahlers.faker.PersonFamilyName
+import ahlers.faker.PersonGivenName
+import org.scalacheck.Arbitrary
 
 /**
  * @author <a href="mailto:michael@ahlers.consulting">Michael Ahlers</a>
