+++ modules/scalacheck/src/main/scala/ahlers/faker/scalacheck/name/implicits/package.scala
@@ -1,0 +1,3 @@
+package ahlers.faker.scalacheck.name
+
+package object implicits extends NameArbitraryInstances
