+++ modules/scalacheck/src/test/scala/ahlers/faker/scalacheck/name/NameSpec.scala
@@ -1,0 +1,16 @@
+package ahlers.faker.scalacheck.name
+
+import org.scalatest.wordspec.AnyWordSpecLike
+import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
+
+class NameSpec extends AnyWordSpecLike {
+  "Person given names" in {
+    import ScalaCheckPropertyChecks._
+    forAll(genPersonGivenName) { _ => }
+  }
+
+  "Person family names" in {
+    import ScalaCheckPropertyChecks._
+    forAll(genPersonFamilyName) { _ => }
+  }
+}
